requirements.txt
# Python version requirement: 3.11+

# Core dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Image processing
opencv-python-headless>=4.8.1.78
numpy>=1.26.3

# OCR
easyocr>=1.7.2
pytesseract>=0.3.10
Pillow>=10.2.0

# NLP/Text processing
python-dateutil>=2.8.2

# Excel support
openpyxl>=3.1.2
pandas>=2.0.0
